requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.8.0rc2
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.115.5
fastapi-cli==0.0.5
flatbuffers==24.3.25
font-amatic-sc==0.0.4
font-fredoka-one==0.0.4
fonts==0.0.3
gast==0.6.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.45
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.2.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikeras==0.13.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
